+++ zcash_keys/src/encoding.rs
@@ -131,16 +131,16 @@ impl<P: consensus::Parameters> AddressCodec<P> for TransparentAddress {
 
     fn encode(&self, params: &P) -> String {
         encode_transparent_address(
-            &params.network_type().b58_pubkey_address_prefix(),
-            &params.network_type().b58_script_address_prefix(),
+            &params.b58_pubkey_address_prefix(),
+            &params.b58_script_address_prefix(),
             self,
         )
     }
 
     fn decode(params: &P, address: &str) -> Result<TransparentAddress, TransparentCodecError> {
         decode_transparent_address(
-            &params.network_type().b58_pubkey_address_prefix(),
-            &params.network_type().b58_script_address_prefix(),
+            &params.b58_pubkey_address_prefix(),
+            &params.b58_script_address_prefix(),
             address,
         )
         .map_err(TransparentCodecError::Base58)
@@ -155,11 +155,11 @@ impl<P: consensus::Parameters> AddressCodec<P> for sapling::PaymentAddress {
     type Error = Bech32DecodeError;
 
     fn encode(&self, params: &P) -> String {
-        encode_payment_address(params.network_type().hrp_sapling_payment_address(), self)
+        encode_payment_address(params.hrp_sapling_payment_address(), self)
     }
 
     fn decode(params: &P, address: &str) -> Result<Self, Bech32DecodeError> {
-        decode_payment_address(params.network_type().hrp_sapling_payment_address(), address)
+        decode_payment_address(params.hrp_sapling_payment_address(), address)
     }
 }
 
@@ -299,7 +299,7 @@ pub fn encode_payment_address_p<P: consensus::Parameters>(
     params: &P,
     addr: &sapling::PaymentAddress,
 ) -> String {
-    encode_payment_address(params.network_type().hrp_sapling_payment_address(), addr)
+    encode_payment_address(params.hrp_sapling_payment_address(), addr)
 }
 
 /// Decodes a [`PaymentAddress`] from a Bech32-encoded string.
@@ -326,7 +326,7 @@ pub fn encode_payment_address_p<P: consensus::Parameters>(
 ///
 /// assert_eq!(
 ///     decode_payment_address(
-///         TEST_NETWORK.network_type().hrp_sapling_payment_address(),
+///         TEST_NETWORK.hrp_sapling_payment_address(),
 ///         "ztestsapling1qqqqqqqqqqqqqqqqqqcguyvaw2vjk4sdyeg0lc970u659lvhqq7t0np6hlup5lusxle75ss7jnk",
 ///     ),
 ///     Ok(pa),
@@ -364,8 +364,8 @@ pub fn decode_payment_address(
 ///
 /// assert_eq!(
 ///     encode_transparent_address(
-///         &TEST_NETWORK.network_type().b58_pubkey_address_prefix(),
-///         &TEST_NETWORK.network_type().b58_script_address_prefix(),
+///         &TEST_NETWORK.b58_pubkey_address_prefix(),
+///         &TEST_NETWORK.b58_script_address_prefix(),
 ///         &TransparentAddress::PublicKeyHash([0; 20]),
 ///     ),
 ///     "tm9iMLAuYMzJ6jtFLcA7rzUmfreGuKvr7Ma",
@@ -373,8 +373,8 @@ pub fn decode_payment_address(
 ///
 /// assert_eq!(
 ///     encode_transparent_address(
-///         &TEST_NETWORK.network_type().b58_pubkey_address_prefix(),
-///         &TEST_NETWORK.network_type().b58_script_address_prefix(),
+///         &TEST_NETWORK.b58_pubkey_address_prefix(),
+///         &TEST_NETWORK.b58_script_address_prefix(),
 ///         &TransparentAddress::ScriptHash([0; 20]),
 ///     ),
 ///     "t26YoyZ1iPgiMEWL4zGUm74eVWfhyDMXzY2",
@@ -411,8 +411,8 @@ pub fn encode_transparent_address_p<P: consensus::Parameters>(
     addr: &TransparentAddress,
 ) -> String {
     encode_transparent_address(
-        &params.network_type().b58_pubkey_address_prefix(),
-        &params.network_type().b58_script_address_prefix(),
+        &params.b58_pubkey_address_prefix(),
+        &params.b58_script_address_prefix(),
         addr,
     )
 }
@@ -432,8 +432,8 @@ pub fn encode_transparent_address_p<P: consensus::Parameters>(
 ///
 /// assert_eq!(
 ///     decode_transparent_address(
-///         &TEST_NETWORK.network_type().b58_pubkey_address_prefix(),
-///         &TEST_NETWORK.network_type().b58_script_address_prefix(),
+///         &TEST_NETWORK.b58_pubkey_address_prefix(),
+///         &TEST_NETWORK.b58_script_address_prefix(),
 ///         "tm9iMLAuYMzJ6jtFLcA7rzUmfreGuKvr7Ma",
 ///     ),
 ///     Ok(Some(TransparentAddress::PublicKeyHash([0; 20]))),
@@ -441,8 +441,8 @@ pub fn encode_transparent_address_p<P: consensus::Parameters>(
 ///
 /// assert_eq!(
 ///     decode_transparent_address(
-///         &TEST_NETWORK.network_type().b58_pubkey_address_prefix(),
-///         &TEST_NETWORK.network_type().b58_script_address_prefix(),
+///         &TEST_NETWORK.b58_pubkey_address_prefix(),
+///         &TEST_NETWORK.b58_script_address_prefix(),
 ///         "t26YoyZ1iPgiMEWL4zGUm74eVWfhyDMXzY2",
 ///     ),
 ///     Ok(Some(TransparentAddress::ScriptHash([0; 20]))),